example-apps/celery-jobs/pythonApps/requirements.txt
pandas==2.2.1
dash==2.15.0
dash-bootstrap-components==1.5.0
dash-mantine-components==0.12.1
dash_ag_grid==31.0.1
girder-client==3.2.2
tqdm==4.66.2
pillow==10.2.0
fastapi==0.104.1
celery==5.3.6
flower==2.0.1
opencv-python==4.9.0.80